Extract account-decoder-client-types (#2872)
* extract account-decoder-client-types and convert UiAccount::encode to a free func encode_ui_account

* rip out more types and replace some usage in rpc-client and rpc-client-api

* rip out more types

* remove remaining account-decoder usage from rpc-client

* remove account-decoder from rpc-client-api and transaction-status-client-types

* make zstd optional

* fix import

* ignore some lint

* update some conversions

* missing import

* missing import

* unused imports

* fmt

* don't feature-gate any variants of UiAccountEncoding

* fmt

* update newer code that used UiAccount::encode

* remove sdk dep from account-decoder-client-types

* add doc_auto_cfg like in #3121
