+++ account-decoder/src/lib.rs
@@ -17,161 +17,86 @@ pub mod parse_token_extension;
 pub mod parse_vote;
 pub mod validator_info;
 
+pub use solana_account_decoder_client_types::{
+    UiAccount, UiAccountData, UiAccountEncoding, UiDataSliceConfig,
+};
 use {
-    crate::parse_account_data::{parse_account_data_v2, AccountAdditionalDataV2, ParsedAccount},
+    crate::parse_account_data::{parse_account_data_v2, AccountAdditionalDataV2},
     base64::{prelude::BASE64_STANDARD, Engine},
-    solana_sdk::{
-        account::{ReadableAccount, WritableAccount},
-        clock::Epoch,
-        fee_calculator::FeeCalculator,
-        pubkey::Pubkey,
-    },
-    std::{
-        io::{Read, Write},
-        str::FromStr,
-    },
+    solana_sdk::{account::ReadableAccount, fee_calculator::FeeCalculator, pubkey::Pubkey},
+    std::io::Write,
 };
 
 pub type StringAmount = String;
 pub type StringDecimals = String;
 pub const MAX_BASE58_BYTES: usize = 128;
 
-/// A duplicate representation of an Account for pretty JSON serialization
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct UiAccount {
-    pub lamports: u64,
-    pub data: UiAccountData,
-    pub owner: String,
-    pub executable: bool,
-    pub rent_epoch: Epoch,
-    pub space: Option<u64>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase", untagged)]
-pub enum UiAccountData {
-    LegacyBinary(String), // Legacy. Retained for RPC backwards compatibility
-    Json(ParsedAccount),
-    Binary(String, UiAccountEncoding),
-}
-
-impl UiAccountData {
-    /// Returns decoded account data in binary format if possible
-    pub fn decode(&self) -> Option<Vec<u8>> {
-        match self {
-            UiAccountData::Json(_) => None,
-            UiAccountData::LegacyBinary(blob) => bs58::decode(blob).into_vec().ok(),
-            UiAccountData::Binary(blob, encoding) => match encoding {
-                UiAccountEncoding::Base58 => bs58::decode(blob).into_vec().ok(),
-                UiAccountEncoding::Base64 => BASE64_STANDARD.decode(blob).ok(),
-                UiAccountEncoding::Base64Zstd => {
-                    BASE64_STANDARD.decode(blob).ok().and_then(|zstd_data| {
-                        let mut data = vec![];
-                        zstd::stream::read::Decoder::new(zstd_data.as_slice())
-                            .and_then(|mut reader| reader.read_to_end(&mut data))
-                            .map(|_| data)
-                            .ok()
-                    })
-                }
-                UiAccountEncoding::Binary | UiAccountEncoding::JsonParsed => None,
-            },
-        }
+fn encode_bs58<T: ReadableAccount>(
+    account: &T,
+    data_slice_config: Option<UiDataSliceConfig>,
+) -> String {
+    let slice = slice_data(account.data(), data_slice_config);
+    if slice.len() <= MAX_BASE58_BYTES {
+        bs58::encode(slice).into_string()
+    } else {
+        "error: data too large for bs58 encoding".to_string()
     }
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub enum UiAccountEncoding {
-    Binary, // Legacy. Retained for RPC backwards compatibility
-    Base58,
-    Base64,
-    JsonParsed,
-    #[serde(rename = "base64+zstd")]
-    Base64Zstd,
-}
-
-impl UiAccount {
-    fn encode_bs58<T: ReadableAccount>(
-        account: &T,
-        data_slice_config: Option<UiDataSliceConfig>,
-    ) -> String {
-        let slice = slice_data(account.data(), data_slice_config);
-        if slice.len() <= MAX_BASE58_BYTES {
-            bs58::encode(slice).into_string()
-        } else {
-            "error: data too large for bs58 encoding".to_string()
+pub fn encode_ui_account<T: ReadableAccount>(
+    pubkey: &Pubkey,
+    account: &T,
+    encoding: UiAccountEncoding,
+    additional_data: Option<AccountAdditionalDataV2>,
+    data_slice_config: Option<UiDataSliceConfig>,
+) -> UiAccount {
+    let space = account.data().len();
+    let data = match encoding {
+        UiAccountEncoding::Binary => {
+            let data = encode_bs58(account, data_slice_config);
+            UiAccountData::LegacyBinary(data)
         }
-    }
-
-    pub fn encode<T: ReadableAccount>(
-        pubkey: &Pubkey,
-        account: &T,
-        encoding: UiAccountEncoding,
-        additional_data: Option<AccountAdditionalDataV2>,
-        data_slice_config: Option<UiDataSliceConfig>,
-    ) -> Self {
-        let space = account.data().len();
-        let data = match encoding {
-            UiAccountEncoding::Binary => {
-                let data = Self::encode_bs58(account, data_slice_config);
-                UiAccountData::LegacyBinary(data)
-            }
-            UiAccountEncoding::Base58 => {
-                let data = Self::encode_bs58(account, data_slice_config);
-                UiAccountData::Binary(data, encoding)
-            }
-            UiAccountEncoding::Base64 => UiAccountData::Binary(
-                BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
-                encoding,
-            ),
-            UiAccountEncoding::Base64Zstd => {
-                let mut encoder = zstd::stream::write::Encoder::new(Vec::new(), 0).unwrap();
-                match encoder
-                    .write_all(slice_data(account.data(), data_slice_config))
-                    .and_then(|()| encoder.finish())
-                {
-                    Ok(zstd_data) => {
-                        UiAccountData::Binary(BASE64_STANDARD.encode(zstd_data), encoding)
-                    }
-                    Err(_) => UiAccountData::Binary(
-                        BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
-                        UiAccountEncoding::Base64,
-                    ),
-                }
-            }
-            UiAccountEncoding::JsonParsed => {
-                if let Ok(parsed_data) =
-                    parse_account_data_v2(pubkey, account.owner(), account.data(), additional_data)
-                {
-                    UiAccountData::Json(parsed_data)
-                } else {
-                    UiAccountData::Binary(
-                        BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
-                        UiAccountEncoding::Base64,
-                    )
-                }
-            }
-        };
-        UiAccount {
-            lamports: account.lamports(),
-            data,
-            owner: account.owner().to_string(),
-            executable: account.executable(),
-            rent_epoch: account.rent_epoch(),
-            space: Some(space as u64),
+        UiAccountEncoding::Base58 => {
+            let data = encode_bs58(account, data_slice_config);
+            UiAccountData::Binary(data, encoding)
         }
-    }
-
-    pub fn decode<T: WritableAccount>(&self) -> Option<T> {
-        let data = self.data.decode()?;
-        Some(T::create(
-            self.lamports,
-            data,
-            Pubkey::from_str(&self.owner).ok()?,
-            self.executable,
-            self.rent_epoch,
-        ))
+        UiAccountEncoding::Base64 => UiAccountData::Binary(
+            BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
+            encoding,
+        ),
+        UiAccountEncoding::Base64Zstd => {
+            let mut encoder = zstd::stream::write::Encoder::new(Vec::new(), 0).unwrap();
+            match encoder
+                .write_all(slice_data(account.data(), data_slice_config))
+                .and_then(|()| encoder.finish())
+            {
+                Ok(zstd_data) => UiAccountData::Binary(BASE64_STANDARD.encode(zstd_data), encoding),
+                Err(_) => UiAccountData::Binary(
+                    BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
+                    UiAccountEncoding::Base64,
+                ),
+            }
+        }
+        UiAccountEncoding::JsonParsed => {
+            if let Ok(parsed_data) =
+                parse_account_data_v2(pubkey, account.owner(), account.data(), additional_data)
+            {
+                UiAccountData::Json(parsed_data)
+            } else {
+                UiAccountData::Binary(
+                    BASE64_STANDARD.encode(slice_data(account.data(), data_slice_config)),
+                    UiAccountEncoding::Base64,
+                )
+            }
+        }
+    };
+    UiAccount {
+        lamports: account.lamports(),
+        data,
+        owner: account.owner().to_string(),
+        executable: account.executable(),
+        rent_epoch: account.rent_epoch(),
+        space: Some(space as u64),
     }
 }
 
@@ -195,13 +120,6 @@ impl Default for UiFeeCalculator {
             lamports_per_signature: "0".to_string(),
         }
     }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiDataSliceConfig {
-    pub offset: usize,
-    pub length: usize,
 }
 
 fn slice_data(data: &[u8], data_slice_config: Option<UiDataSliceConfig>) -> &[u8] {
@@ -264,13 +182,13 @@ mod test {
 
         // Whole account
         assert_eq!(
-            UiAccount::encode_bs58(&account, None),
+            encode_bs58(&account, None),
             "error: data too large for bs58 encoding"
         );
 
         // Slice of account that's still too large
         assert_eq!(
-            UiAccount::encode_bs58(
+            encode_bs58(
                 &account,
                 Some(UiDataSliceConfig {
                     length: MAX_BASE58_BYTES + 1,
@@ -282,7 +200,7 @@ mod test {
 
         // Slice of account that fits inside `MAX_BASE58_BYTES`
         assert_ne!(
-            UiAccount::encode_bs58(
+            encode_bs58(
                 &account,
                 Some(UiDataSliceConfig {
                     length: MAX_BASE58_BYTES,
@@ -294,7 +212,7 @@ mod test {
 
         // Slice of account that's too large, but whose intersection with the account still fits
         assert_ne!(
-            UiAccount::encode_bs58(
+            encode_bs58(
                 &account,
                 Some(UiDataSliceConfig {
                     length: MAX_BASE58_BYTES + 1,
@@ -307,7 +225,7 @@ mod test {
 
     #[test]
     fn test_base64_zstd() {
-        let encoded_account = UiAccount::encode(
+        let encoded_account = encode_ui_account(
             &Pubkey::default(),
             &AccountSharedData::from(Account {
                 data: vec![0; 1024],
