+++ account-decoder/src/parse_account_data.rs
@@ -1,3 +1,4 @@
+pub use solana_account_decoder_client_types::ParsedAccount;
 use {
     crate::{
         parse_address_lookup_table::parse_address_lookup_table,
@@ -6,7 +7,6 @@ use {
         parse_token::parse_token_v2, parse_vote::parse_vote,
     },
     inflector::Inflector,
-    serde_json::Value,
     solana_sdk::{
         address_lookup_table, clock::UnixTimestamp, instruction::InstructionError, pubkey::Pubkey,
         stake, system_program, sysvar, vote,
@@ -61,14 +61,6 @@ pub enum ParseAccountError {
 
     #[error("Serde json error")]
     SerdeJsonError(#[from] serde_json::error::Error),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct ParsedAccount {
-    pub program: String,
-    pub parsed: Value,
-    pub space: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
