+++ rpc-client-api/Cargo.toml
@@ -20,7 +20,7 @@ semver = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 serde_json = { workspace = true }
-solana-account-decoder = { workspace = true }
+solana-account-decoder-client-types = { workspace = true }
 solana-inline-spl = { workspace = true }
 solana-sdk = { workspace = true }
 solana-transaction-status-client-types = { workspace = true }
