+++ rpc-client/Cargo.toml
@@ -22,7 +22,7 @@ semver = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 serde_json = { workspace = true }
-solana-account-decoder = { workspace = true }
+solana-account-decoder-client-types = { workspace = true, features = ["zstd"] }
 solana-rpc-client-api = { workspace = true }
 solana-sdk = { workspace = true }
 solana-transaction-status-client-types = { workspace = true }
@@ -36,6 +36,7 @@ crossbeam-channel = { workspace = true }
 futures = { workspace = true }
 jsonrpc-core = { workspace = true }
 jsonrpc-http-server = { workspace = true }
+solana-account-decoder = { workspace = true }
 
 [features]
 default = ["spinner"]
