+++ rpc-client/src/mock_sender.rs
@@ -5,7 +5,7 @@ use {
     async_trait::async_trait,
     base64::{prelude::BASE64_STANDARD, Engine},
     serde_json::{json, Number, Value},
-    solana_account_decoder::{UiAccount, UiAccountEncoding},
+    solana_account_decoder_client_types::{UiAccount, UiAccountData, UiAccountEncoding},
     solana_rpc_client_api::{
         client_error::Result,
         config::RpcBlockProductionConfig,
@@ -20,7 +20,6 @@ use {
         },
     },
     solana_sdk::{
-        account::Account,
         clock::{Slot, UnixTimestamp},
         epoch_info::EpochInfo,
         instruction::InstructionError,
@@ -426,23 +425,10 @@ impl RpcSender for MockSender {
             })?,
             "getProgramAccounts" => {
                 let pubkey = Pubkey::from_str(PUBKEY).unwrap();
-                let account = Account {
-                    lamports: 1_000_000,
-                    data: vec![],
-                    owner: pubkey,
-                    executable: false,
-                    rent_epoch: 0,
-                };
                 serde_json::to_value(vec![
                     RpcKeyedAccount {
                         pubkey: PUBKEY.to_string(),
-                        account: UiAccount::encode(
-                            &pubkey,
-                            &account,
-                            UiAccountEncoding::Base64,
-                            None,
-                            None,
-                        )
+                        account: mock_encoded_account(&pubkey)
                     }
                 ])?
             },
@@ -453,5 +439,35 @@ impl RpcSender for MockSender {
 
     fn url(&self) -> String {
         format!("MockSender: {}", self.url)
+    }
+}
+
+pub(crate) fn mock_encoded_account(pubkey: &Pubkey) -> UiAccount {
+    UiAccount {
+        lamports: 1_000_000,
+        data: UiAccountData::Binary("".to_string(), UiAccountEncoding::Base64),
+        owner: pubkey.to_string(),
+        executable: false,
+        rent_epoch: 0,
+        space: Some(0),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use {super::*, solana_account_decoder::encode_ui_account, solana_sdk::account::Account};
+
+    #[test]
+    fn test_mock_encoded_account() {
+        let pubkey = Pubkey::from_str(PUBKEY).unwrap();
+        let account = Account {
+            lamports: 1_000_000,
+            data: vec![],
+            owner: pubkey,
+            executable: false,
+            rent_epoch: 0,
+        };
+        let expected = encode_ui_account(&pubkey, &account, UiAccountEncoding::Base64, None, None);
+        assert_eq!(expected, mock_encoded_account(&pubkey));
     }
 }
