+++ rpc/src/rpc_subscriptions.rs
@@ -17,7 +17,9 @@ use {
     itertools::Either,
     rayon::prelude::*,
     serde::Serialize,
-    solana_account_decoder::{parse_token::is_known_spl_token_id, UiAccount, UiAccountEncoding},
+    solana_account_decoder::{
+        encode_ui_account, parse_token::is_known_spl_token_id, UiAccount, UiAccountEncoding,
+    },
     solana_ledger::{blockstore::Blockstore, get_tmp_ledger_path},
     solana_measure::measure::Measure,
     solana_rpc_client_api::response::{
@@ -385,7 +387,7 @@ fn filter_account_result(
         {
             get_parsed_token_account(&bank, &params.pubkey, account, None)
         } else {
-            UiAccount::encode(&params.pubkey, &account, params.encoding, None, None)
+            encode_ui_account(&params.pubkey, &account, params.encoding, None, None)
         }
     });
     (account, last_modified_slot)
@@ -428,7 +430,7 @@ fn filter_program_results(
     } else {
         let accounts = keyed_accounts.map(move |(pubkey, account)| RpcKeyedAccount {
             pubkey: pubkey.to_string(),
-            account: UiAccount::encode(&pubkey, &account, encoding, None, None),
+            account: encode_ui_account(&pubkey, &account, encoding, None, None),
         });
         Either::Right(accounts)
     };
