+++ transaction-status/src/parse_token/extension/default_account_state.rs
@@ -1,5 +1,6 @@
 use {
     super::*,
+    solana_account_decoder::parse_token::convert_account_state,
     spl_token_2022::extension::default_account_state::instruction::{
         decode_instruction, DefaultAccountStateInstruction,
     },
@@ -22,7 +23,7 @@ pub(in crate::parse_token) fn parse_default_account_state_instruction(
                 instruction_type: format!("initialize{instruction_type}"),
                 info: json!({
                     "mint": account_keys[account_indexes[0] as usize].to_string(),
-                    "accountState": UiAccountState::from(account_state),
+                    "accountState": convert_account_state(account_state),
                 }),
             })
         }
@@ -30,7 +31,7 @@ pub(in crate::parse_token) fn parse_default_account_state_instruction(
             check_num_token_accounts(account_indexes, 2)?;
             let mut value = json!({
                 "mint": account_keys[account_indexes[0] as usize].to_string(),
-                "accountState": UiAccountState::from(account_state),
+                "accountState": convert_account_state(account_state),
             });
             let map = value.as_object_mut().unwrap();
             parse_signers(
